moved files to article-specific folder
