+++ core-java-modules/core-java-lang-oop-types-2/src/main/java/com/baeldung/enums/Direction.java
@@ -1,24 +1,10 @@
 package com.baeldung.enums;
 
-import java.util.Random;
-
 /**
  * Represents directions.
  */
 public enum Direction {
     EAST, WEST, SOUTH, NORTH;
-
-    private static final Random PRNG = new Random();
-
-    /**
-     * Generate a random direction.
-     *
-     * @return a random direction
-     */
-    public static Direction randomDirection() {
-        Direction[] directions = values();
-        return directions[PRNG.nextInt(directions.length)];
-    }
 
     /**
      * Finds direction by name.
