+++ core-java-modules/core-java-lang-oop-types-2/src/main/java/com/baeldung/enums/randomenum/RandomEnumGenerator.java
@@ -1,0 +1,17 @@
+package com.baeldung.enums.randomenum;
+
+import java.util.Random;
+
+public class RandomEnumGenerator<T extends Enum<T>> {
+
+    private static final Random PRNG = new Random();
+    private final T[] values;
+
+    public RandomEnumGenerator(Class<T> e) {
+        values = e.getEnumConstants();
+    }
+
+    public T randomEnum() {
+        return values[PRNG.nextInt(values.length)];
+    }
+}
